ChatBot/Local Bot/requirements.txt
python-telegram-bot==13.7 